# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lerank platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  student: "student",
  company_admin: "company_admin",
  super_admin: "super_admin",
} as const;

export interface StudentProfile {
  ieltsScore?: number | null;
  satScore?: number | null;
  toeflScore?: number | null;
  gpa?: number | null;
  gpaScale?: number | null;
  budgetMin?: number | null;
  budgetMax?: number | null;
  educationBudget?: string | null;
  preferredCountries?: string[];
  preferredMajor?: string | null;
  degreeLevel?: string | null;
  additionalNotes?: string | null;
}

export interface User {
  id: number;
  email: string;
  fullName: string;
  role: UserRole;
  avatarUrl?: string | null;
  onboardingCompleted: boolean;
  profile?: StudentProfile | null;
  companyId?: number | null;
  createdAt: string;
}

export interface RegisterRequest {
  email: string;
  /** @minLength 6 */
  password: string;
  fullName: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UpdateProfileRequest {
  ieltsScore?: number | null;
  satScore?: number | null;
  toeflScore?: number | null;
  gpa?: number | null;
  gpaScale?: number | null;
  budgetMin?: number | null;
  budgetMax?: number | null;
  educationBudget?: string | null;
  preferredCountries?: string[];
  preferredMajor?: string | null;
  degreeLevel?: string | null;
  additionalNotes?: string | null;
  onboardingCompleted?: boolean;
}

export interface ConsultantCompany {
  id: number;
  name: string;
  logoUrl?: string | null;
  description?: string | null;
  website?: string | null;
  rating: number;
  studentsHelped: number;
  successRate: number;
  priceMin: number;
  priceMax: number;
  specializedCountries: string[];
  degreeLevels: string[];
  isActive: boolean;
  createdAt: string;
}

export interface CreateCompanyRequest {
  name: string;
  description?: string | null;
  website?: string | null;
  priceMin: number;
  priceMax: number;
  specializedCountries: string[];
  degreeLevels: string[];
}

export interface UpdateCompanyRequest {
  name?: string;
  description?: string | null;
  website?: string | null;
  isActive?: boolean;
  priceMin?: number;
  priceMax?: number;
}

export type ApplicationStatus =
  (typeof ApplicationStatus)[keyof typeof ApplicationStatus];

export const ApplicationStatus = {
  active: "active",
  completed: "completed",
  refunded: "refunded",
  disputed: "disputed",
} as const;

export type ApplicationEscrowStatus =
  (typeof ApplicationEscrowStatus)[keyof typeof ApplicationEscrowStatus];

export const ApplicationEscrowStatus = {
  pending: "pending",
  held: "held",
  released: "released",
  refunded: "refunded",
} as const;

export interface Application {
  id: number;
  studentId: number;
  companyId: number;
  companyName: string;
  serviceName: string;
  status: ApplicationStatus;
  progressPercent: number;
  totalAmount: number;
  escrowStatus: ApplicationEscrowStatus;
  createdAt: string;
  updatedAt: string;
}

export type MilestoneStatus =
  (typeof MilestoneStatus)[keyof typeof MilestoneStatus];

export const MilestoneStatus = {
  not_started: "not_started",
  in_progress: "in_progress",
  under_review: "under_review",
  completed: "completed",
} as const;

export interface Milestone {
  id: number;
  applicationId: number;
  name: string;
  description?: string | null;
  status: MilestoneStatus;
  orderIndex: number;
  updatedAt: string;
}

export type ApplicationDetail = Application & {
  milestones: Milestone[];
};

export interface CreateApplicationRequest {
  companyId: number;
  serviceName: string;
  totalAmount: number;
  templateId?: number | null;
}

export type UpdateMilestoneRequestStatus =
  (typeof UpdateMilestoneRequestStatus)[keyof typeof UpdateMilestoneRequestStatus];

export const UpdateMilestoneRequestStatus = {
  not_started: "not_started",
  in_progress: "in_progress",
  under_review: "under_review",
  completed: "completed",
} as const;

export interface UpdateMilestoneRequest {
  status: UpdateMilestoneRequestStatus;
  description?: string | null;
}

export type TransactionStatus =
  (typeof TransactionStatus)[keyof typeof TransactionStatus];

export const TransactionStatus = {
  pending: "pending",
  held: "held",
  released: "released",
  refunded: "refunded",
  disputed: "disputed",
} as const;

export interface Transaction {
  id: number;
  applicationId: number;
  studentName: string;
  companyName: string;
  amount: number;
  platformFee: number;
  escrowFee: number;
  status: TransactionStatus;
  createdAt: string;
}

export interface DashboardStats {
  activeApplications: number;
  milestonesCompletedThisWeek: number;
  moneyInEscrow: number;
  moneyReleased: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  milestone_update: "milestone_update",
  payment: "payment",
  application: "application",
} as const;

export interface ActivityItem {
  id: number;
  message: string;
  timestamp: string;
  type: ActivityItemType;
}

export interface AdminAnalytics {
  totalStudents: number;
  totalCompanies: number;
  totalRevenue: number;
  activeApplications: number;
  pendingTransactions: number;
  completedTransactions: number;
  refundedTransactions: number;
}

export type StudentWithApplicationEscrowStatus =
  (typeof StudentWithApplicationEscrowStatus)[keyof typeof StudentWithApplicationEscrowStatus];

export const StudentWithApplicationEscrowStatus = {
  pending: "pending",
  held: "held",
  released: "released",
  refunded: "refunded",
} as const;

export interface StudentWithApplication {
  studentId: number;
  studentName: string;
  email: string;
  applicationId: number;
  serviceName: string;
  progressPercent: number;
  escrowStatus: StudentWithApplicationEscrowStatus;
  totalAmount: number;
}

export type MilestoneTemplateMilestonesItem = {
  name: string;
  orderIndex: number;
};

export interface MilestoneTemplate {
  id: number;
  companyId: number;
  name: string;
  milestones: MilestoneTemplateMilestonesItem[];
  createdAt: string;
}

export type CreateMilestoneTemplateRequestMilestonesItem = {
  name: string;
  orderIndex: number;
};

export interface CreateMilestoneTemplateRequest {
  name: string;
  milestones: CreateMilestoneTemplateRequestMilestonesItem[];
}

export type ListConsultantsParams = {
  country?: string;
  minRating?: number;
  maxPrice?: number;
  degreeLevel?: ListConsultantsDegreeLevel;
  sortBy?: ListConsultantsSortBy;
};

export type ListConsultantsDegreeLevel =
  (typeof ListConsultantsDegreeLevel)[keyof typeof ListConsultantsDegreeLevel];

export const ListConsultantsDegreeLevel = {
  bachelor: "bachelor",
  master: "master",
  phd: "phd",
} as const;

export type ListConsultantsSortBy =
  (typeof ListConsultantsSortBy)[keyof typeof ListConsultantsSortBy];

export const ListConsultantsSortBy = {
  bestMatch: "bestMatch",
  priceLow: "priceLow",
  rating: "rating",
  mostStudents: "mostStudents",
} as const;