# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Lerank platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /applications/{id}/milestones:
    get:
      operationId: getApplicationMilestones
      tags: [milestones]
      summary: Get milestones for an application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Milestones list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Milestone"
  /milestones/{id}:
    put:
      operationId: updateMilestone
      tags: [milestones]
      summary: Update milestone status (company admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMilestoneRequest"
      responses:
        "200":
          description: Milestone updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
...
components:
  schemas:
...
    Milestone:
      type: object
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [not_started, in_progress, under_review, completed]
        orderIndex:
          type: integer
        updatedAt:
          type: string
          format: date-time
      required: [id, applicationId, name, status, orderIndex, updatedAt]
    UpdateMilestoneRequest:
      type: object
      properties:
        status:
          type: string
          enum: [not_started, in_progress, under_review, completed]
        description:
          type: string
          nullable: true
      required: [status]
...