# lib/api-zod/src/generated/types/updateMilestoneRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lerank platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateMilestoneRequestStatus } from "./updateMilestoneRequestStatus";

export interface UpdateMilestoneRequest {
  status: UpdateMilestoneRequestStatus;
  description?: string | null;
}